google-a2a/requirements.txt
a2a-sdk[http-server]>=0.3.0
openai>=1.75.0
python-dotenv>=1.0.1
uvicorn>=0.34.0
